requirements.txt
nilearn
numpy
PyYAML
scikit_learn
scipy
torch
torch_geometric
tqdm
